backend/requirements.txt
fastapi==0.115.12
langchain==0.3.23
langchain_community==0.3.21
langchain_core==0.3.52
langchain_ollama==0.3.1
langchain_openai==0.3.13
langchain_qdrant==0.2.0
langchain_text_splitters==0.3.8
pydantic==2.11.3
python-dotenv==1.1.0
qdrant_client==1.13.3
tqdm==4.67.1
uvicorn==0.34.1
